🚚MainService.js에 프로필이미지, 비디오시간, 조회수 코드를 utils폴더에 따로 생성
